moving masters to armasters.py
