+++ pypit/arsave.py
@@ -125,91 +125,7 @@ def save_extraction(slf, sciext, scidx, scierr=None, filename="temp.fits", frame
     return
 
 
-def save_master(slf, data, filename="temp.fits", frametype="<None>", ind=[],
-                        extensions=None, keywds=None, names=None):
-    """ Wrapper to core_save_master
-    Will be Deprecated
 
-    Parameters
-    ----------
-    slf
-    data
-    filename
-    frametype
-    ind
-    extensions
-    keywds
-    names
-
-    Returns
-    -------
-
-    """
-    if len(ind) > 0:
-        raw_files=slf._fitsdict['filename']
-    else:
-        raw_files=None
-    core_save_master(data, filename=filename, frametype=frametype, ind=ind,
-                     extensions=extensions, keywds=keywds, names=names,
-                     raw_files=raw_files)
-
-
-def core_save_master(data, filename="temp.fits", frametype="<None>", ind=[],
-                extensions=None, keywds=None, names=None, raw_files=None,
-                     overwrite=True):
-    """ Core function to write a MasterFrame
-    Parameters
-    ----------
-    data : ndarray
-    filename : str (optional)
-    frametype : str (optional)
-    ind : list (optional)
-    extensions : list, optional
-      Additional data images to write
-    names : list, optional
-      Names of the extensions
-    keywds : Additional keywords for the Header
-    raw_files : list or ndarray
-      Names of the raw files used to generate the image
-
-    Returns
-    -------
-
-    """
-    # Check for existing
-    if os.path.exists(filename) and (not overwrite):
-        msgs.warn("This file already exists.  Use overwrite=True to overwrite it")
-        return
-    #
-    msgs.info("Saving master {0:s} frame as:".format(frametype)+msgs.newline()+filename)
-    hdu = fits.PrimaryHDU(data)
-    hlist = [hdu]
-    # Extensions
-    if extensions is not None:
-        for kk,exten in enumerate(extensions):
-            hdu = fits.ImageHDU(exten)
-            if names is not None:
-                hdu.name = names[kk]
-            hlist.append(hdu)
-    # HDU list
-    hdulist = fits.HDUList(hlist)
-    # Header
-    msgs.info("Writing header information")
-    for i in range(len(ind)):
-        hdrname = "FRAME{0:03d}".format(i+1)
-        if raw_files is not None:
-            hdulist[0].header[hdrname] = (raw_files[ind[i]], 'PYPIT: File used to generate Master {0:s}'.format(frametype))
-    hdulist[0].header["FRAMETYP"] = (frametype, 'PYPIT: Master calibration frame type')
-    if keywds is not None:
-        for key in keywds.keys():
-            hdulist[0].header[key] = keywds[key]
-    # Write the file to disk
-    if os.path.exists(filename):
-        msgs.warn("Overwriting file:"+msgs.newline()+filename)
-
-    hdulist.writeto(filename, overwrite=True)
-    msgs.info("Master {0:s} frame saved successfully:".format(frametype)+msgs.newline()+filename)
-    return
 
 
 def save_ordloc(slf, fname):
